finished KaleKing and windups

+++ src/pizzatales/FlamerFlame.java
@@ -6,7 +6,7 @@ public class FlamerFlame extends Projectile {
 	
 	public static Image bulletsprite;
 	
-	public FlamerFlame(int startX, int startY, float vectorX, float vectorY, int speed, float dmg, int range) {
+	public FlamerFlame(int startX, int startY, float vectorX, float vectorY, float speed, float dmg, int range) {
 		super(startX, startY, vectorX, vectorY, speed, dmg, 30, 15, range);
 	}
